Move mintable tokens to utils and add minter
